create utils file with reused code
